feat: separação das rotas de agenda
