+++ server/data/agenda.json
@@ -26,8 +26,8 @@
     "title": "Agenda aulas3",
     "url": "https://meet.google.com/ywa-widc-fsf",
     "duration": "01:00",
-    "start": "2022-06-01",
-    "end": "2022-06-30",
+    "start": "2022-07-01",
+    "end": "2022-07-30",
     "businessHours": [{
         "daysOfWeek": [2, 3],
         "startTime": "08:00",
