+++ server/routes/account.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 const controllerAccount = require('../controllers/AccountController')
-
+const controllerAgenda = require('../controllers/AgendaController')
 
 router.get('/:userId', controllerAccount.calendar)
 
@@ -12,12 +12,12 @@ router.get('/:userId/minha-conta', controllerAccount.account)
 router.get('/:userId/minha-conta/editar',controllerAccount.editAccount)
 // router.get('/assinatura/editar', controllerAccount.editSignature)
 
-router.get('/:userId/criar-agenda', controllerAccount.addAgenda)
-router.post('/:userId/criar-agenda', controllerAccount.createAgenda)
-router.get('/:userId/editar-agenda', controllerAccount.editAgenda)
-router.put('/:userId/editar-agenda', controllerAccount.updateAgenda)
-router.get('/:userId/apagar-agenda', controllerAccount.removeAgenda)
-router.delete('/:userId/apagar-agenda', controllerAccount.deleteAgenda)
+router.get('/:userId/criar-agenda', controllerAgenda.addAgenda)
+router.post('/:userId/criar-agenda', controllerAgenda.createAgenda)
+router.get('/:userId/editar-agenda', controllerAgenda.editAgenda)
+router.put('/:userId/editar-agenda', controllerAgenda.updateAgenda)
+router.get('/:userId/apagar-agenda', controllerAgenda.removeAgenda)
+router.delete('/:userId/apagar-agenda', controllerAgenda.deleteAgenda)
 
 router.get('/json', controllerAccount.events)
 router.get('/json/:id', controllerAccount.eventsId)
