+++ server/routes/agendas.js
@@ -1,0 +1,10 @@
+var express = require('express');
+var router = express.Router();
+const controllerAgendas = require('../controllers/AgendaController')
+
+
+router.get('/:slug', controllerAgendas.agendas)
+router.get('/:slug/:id', controllerAgendas.showAgenda)
+
+
+module.exports = router
