+++ server/views/agendas.ejs
@@ -21,7 +21,7 @@
     </header>
     <main>
         <section>
-            <img src="/images/card-img4.jpg" width="100%">
+            <img src="/images/card-img4.jpg" height="600px" width="100%" >
         </section>
         <section class="agendas">
             <div>
